kernel/fork: Boost only for 50ms on zygote fork
* boosting to max for 1s on every zygote fork causes unwanted battery drain

* from sultanxda:
In order to boost during the period from when a new app is forked to the
time that it is added to the top-app cpuset (which is a signal to
subsequently boost again), boost for a short amount of time when zygote
forks. Not all processes forked from zygote are top apps, hence the
decoupled boosting setup.